chore: Combined transfer_controller into projects_controller
